Fix premature Elemental Invasion boss despawn.
Fixes https://github.com/slowtorta/turtlewow-bug-tracker/issues/255

+++ src/game/HardcodedEvents.h
@@ -55,8 +55,8 @@ struct ElementalInvasion : WorldEvent
 
 private:
     void StartLocalInvasion(uint8 index, uint32 stage);
-    void StartLocalBoss(uint8 index, uint32 stage, uint8 delay);
-    void StopLocalInvasion(uint8 index, uint32 stage, uint8 delay);
+    void StartLocalBoss(uint8 index, uint32 stage, uint32 delay);
+    void StopLocalInvasion(uint8 index, uint32 stage, uint32 delay);
     void ResetThings();
 };
